Move methods to a util fn
